requirements.txt
python-dotenv~=0.19.2
sanic~=22.6.2
pymongo~=4.2.0
aiohttp~=3.7.4